requirements.txt
streamlit
pandas
fuzzywuzzy
python-docx
pdfplumber
openpyxl
